Creacion de EditClienteManager y retoques en front
