+++ app/src/main/java/com/example/proyectointegradordam/RegistroSocio.kt
@@ -63,9 +63,6 @@ class RegistroSocio : BaseActivity() {
                     Toast.makeText(this, "Error al registrar socio", Toast.LENGTH_SHORT).show()
                 }
             }
-
-        }
-
-
         }
     }
+}
