+++ app/src/main/java/com/example/proyectointegradordam/database/ClubDeportivoDBHelper.kt
@@ -92,7 +92,6 @@ class clubDeportivoDBHelper(context: Context) : SQLiteOpenHelper(context, DATABA
     }
 
 
-
     override fun onUpgrade(db: SQLiteDatabase, oldVersion: Int, newVersion: Int) {
         db.execSQL("DROP TABLE IF EXISTS credito_actividades")
         db.execSQL("DROP TABLE IF EXISTS usuario")
@@ -103,66 +102,5 @@ class clubDeportivoDBHelper(context: Context) : SQLiteOpenHelper(context, DATABA
         onCreate(db)
     }
 
-
-
-    fun buscarClientePorNombre(texto: String): List<Cliente>{
-        val db = readableDatabase
-        val cursor = db.rawQuery(
-            "SELECT * FROM cliente WHERE nombre LIKE ? OR apellido LIKE ?",
-            arrayOf("%$texto%", "%$texto%")
-        )
-        val lista = mutableListOf<Cliente>()
-        if(cursor.moveToFirst()){
-            do{
-                lista.add(
-                    Cliente(
-                        cursor.getInt(cursor.getColumnIndexOrThrow("id_cliente")),
-                        cursor.getString(cursor.getColumnIndexOrThrow("nombre")),
-                        cursor.getString(cursor.getColumnIndexOrThrow("apellido")),
-                        cursor.getString(cursor.getColumnIndexOrThrow("email")),
-                        cursor.getString(cursor.getColumnIndexOrThrow("telefono"))
-                    )
-                )
-            } while (cursor.moveToNext())
-        }
-        cursor.close()
-        return lista
-
-    }
-
-    fun actualizarDatosCliente(
-        id: Int,
-        email: String,
-        telefono: String
-    ): Int{
-        val db = writableDatabase
-        val values = ContentValues().apply {
-            put("email", email)
-            put("telefono", telefono)
-        }
-        return db.update("cliente", values, "id_cliente = ?", arrayOf(id.toString()))
-
-    }
-
-    fun obtenerTodosLosClientes(): List<Cliente> {
-        val db = readableDatabase
-        val cursor = db.rawQuery("SELECT * FROM cliente", null)
-        val lista = mutableListOf<Cliente>()
-        if (cursor.moveToFirst()) {
-            do {
-                lista.add(
-                    Cliente(
-                        cursor.getInt(cursor.getColumnIndexOrThrow("id_cliente")),
-                        cursor.getString(cursor.getColumnIndexOrThrow("nombre")),
-                        cursor.getString(cursor.getColumnIndexOrThrow("apellido")),
-                        cursor.getString(cursor.getColumnIndexOrThrow("email")),
-                        cursor.getString(cursor.getColumnIndexOrThrow("telefono"))
-                    )
-                )
-            } while (cursor.moveToNext())
-        }
-        cursor.close()
-        return lista
-    }
 
 }
