+++ app/src/main/java/com/example/proyectointegradordam/view/ActividadesActivity.kt
@@ -28,6 +28,7 @@ import com.example.proyectointegradordam.managers.ActivitiesManager
 import com.example.proyectointegradordam.managers.InscripcionManager
 import com.example.proyectointegradordam.models.Activities
 import com.example.proyectointegradordam.models.Cliente
+import com.example.proyectointegradordam.managers.EditClienteManager
 
 class ActividadesActivity : BaseActivity(), ActivityAdapter.OnActividadActualizadaListener {
 
@@ -70,12 +71,13 @@ class ActividadesActivity : BaseActivity(), ActivityAdapter.OnActividadActualiza
 
         val dbHelper = clubDeportivoDBHelper(this)
         val inscripcionManager = InscripcionManager(this)
+        val editClienteManager = EditClienteManager(this)
 
         var clienteSeleccionado: Cliente? = null
         var actividadSeleccionada: Activities? = null
 
         // Obtener todos los clientes
-        val todosLosClientes = dbHelper.obtenerTodosLosClientes()
+        val todosLosClientes = editClienteManager.obtenerTodosLosClientes()
         val spinnerClientes = modalBinding.root.findViewById<Spinner>(R.id.spinnerClientes)
         val spinnerActividades = modalBinding.root.findViewById<Spinner>(R.id.spinnerActividades)
