+++ app/src/main/res/drawable/recycler_vista.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="#FFFFFF"/>
+    <stroke android:width="2dp" android:color="@color/grey"/>
+    <corners android:radius="8dp"/>
+
+</shape>
