+++ app/src/main/res/layout/item_cliente.xml
@@ -7,5 +7,6 @@ xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_height="wrap_content"
     android:padding="16dp"
     android:textSize="16sp"
+    android:fontFamily="@font/comfortaa_regular"
     android:textColor="@color/black"/>
