+++ app/src/main/res/values/colors.xml
@@ -2,6 +2,7 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+    <color name="grey">#555555</color>
 
     <color name="bc_backgroundButton">#2E6200</color>
